コンフリクト解消後の調整

+++ src/Order.java
@@ -1,4 +1,5 @@
 import java.util.Scanner;
+import java.math.BigDecimal;
 
 public class Order {
   //::::::::::
@@ -134,7 +135,7 @@ public class Order {
         this.slipNo.equals("none")) {
       System.out.print(">> 運賃計算が可能です 計算しますか？(1:する / 0:しない) -> ");
       settingInt = settingIntScanner.nextInt();
-      this.loadFare = 999; // 運賃計算メソッドが治るまでコレ
+      this.loadFare = (BigDecimal)999; // 運賃計算メソッドが治るまでコレ
     }
     System.out.println(">> 受注情報入力完了");
     this.displayInfo();
